# Conflicts:
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/controller/EventController.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/controller/UserController.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/Location.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/entity/User.java
#	backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/service/impl/PerformanceServiceImpl.java

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/repository/EventRepository.java
@@ -1,10 +1,9 @@
 package at.ac.tuwien.sepm.groupphase.backend.repository;
 
 import at.ac.tuwien.sepm.groupphase.backend.entity.Event;
-import at.ac.tuwien.sepm.groupphase.backend.entity.Performance;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event, Integer> {
+public interface EventRepository extends JpaRepository<Event, Long> {
 }
